// Sensalon_back/src/interfaces/Salon.ts
export interface Salon {
    iIdsalon: string;  
    iFIdUser: string;  
    vcsalonname: string;  
    vccellphone?: string;  
    vcemail?: string;  
    vcaddress?: string;  
    dtopeningtime?: string; 
    dtdeparturtime?: string;  
    vcservices?: string;  
    vclogo?: string;
    dtcreation?: Date;
}